Upgraded to NHibernate 1.2 beta 2. Using newer Puzzle and Generics.

+++ SnCore.Web/TagWordsView.aspx.cs
@@ -105,7 +105,7 @@ public partial class TagWordsView : AccountPersonPage
             }
     }
 
-    public int GetFontSize(int frequency)
+    public long GetFontSize(int frequency)
     {
         int frequencyDelta = MaxFrequency - MinFrequency;
         if (frequencyDelta <= 0)
